refactor: move auto-mnemonic code to separate file and deduplicate remove_mnemonic_marker code
